Add a little documentation for backcompat and reorganize for tests
