+++ bids-validator/src/compat/README.md
@@ -1,0 +1,1 @@
+Support files for running the JavaScript validator from TypeScript and/or Deno. This is used to run legacy validator issues after schema based validation.
