+++ bids-validator/src/compat/adapter-file.ts
@@ -1,0 +1,20 @@
+import { relative, resolve } from '../deps/path.ts'
+import { BIDSFile } from '../files/filetree.ts'
+
+/**
+ * File class modeled off browser File for validator
+ */
+export class AdapterFile {
+  name: string // Filename
+  path: string // Absolute path
+  relativePath: string // Dataset relative path (prefixed /)
+  webkitRelativePath: string // Duplicate name for relativePath
+  constructor(rootPath: string, file: BIDSFile) {
+    this.name = file.name
+    // JS validator expects dataset-dir/contents filenames
+    const relativePath = relative(rootPath, file.path)
+    this.relativePath = `/${relativePath}`
+    this.webkitRelativePath = this.relativePath
+    this.path = resolve(rootPath, file.path)
+  }
+}
